Left Indicator width not updating on setter
When the left indicator width is set, it does not redraw the view with
the new width. Even if the view is forced to redraw (e.g. change the
frame or call setNeedsLayout), layoutSubviews will still use the old
cached value not the new value. The only way to force the new left
indicator width to take effect is by calling setText or changing the
text in the text field. I have created custom setters that will
correctly redraw the view when the indicator width is changed.